pyproject.toml
[tool.poetry]
name = "devcontainer-glue-template"
version = "0.1.0"
description = "Dev container template for AWS Glue jobs"
authors = ["wtfzambo <zambelli.federico@gmail.com>"]
readme = "README.md"
# packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = "^3.10"
boto3 = "^1.26.97"

[tool.poetry.group.dev.dependencies]
chispa = "^0.9.2"
flake8 = "^6.0.0"
flake8-black = "^0.3.6"
black = "^23.1.0"
isort = "^5.12.0"
pytest = "^7.2.2"
python-dotenv = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
force_single_line = true
lines_after_imports = 2
